+++ tests/testthat/test-mod-cluster.r
@@ -26,4 +26,15 @@ test_that("Testes de clusterizacao", {
 
     gg <- plot(compac, clust, print = FALSE)
     expect_equal(class(gg), c("plotly", "htmlwidget"))
+
+    # HCLUST
+
+    clust <- clusthierarq(compac, 5)
+
+    expect_equal(class(clust), c("hclust_aug", "hclust"))
+    expect_snapshot_value(getclustclass(clust), style = "deparse")
+    expect_snapshot_value(getclustmeans(clust), style = "deparse")
+
+    gg <- plot(compac, clust, print = FALSE)
+    expect_equal(class(gg), c("plotly", "htmlwidget"))
 })